pyproject.toml
[tool.poetry]
name = "marl-ripped-qmix"
version = "0.1.0"
description = ""
authors = ["Sithael <chojnacki.jamie@gmail.com>"]
license = "MIT license"
readme = "README.md"
packages = [{include = "qmix"}]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.25.1"
torch = "^2.0.1"
pettingzoo = "^1.23.1"
pyyaml = "^6.0.1"
pytest = "^7.4.0"
black = "^23.7.0"
poethepoet = "^0.21.1"
pre-commit = "^3.3.3"
click = "^8.1.6"
toml = "^0.10.2"
box2d-py = "^2.3.8"
pygame = "^2.5.0"
pysc2 = "^4.0.0"
protobuf = "3.20"
smac = {git = "https://github.com/oxwhirl/smac.git"}
mypy = "^1.5.1"
torchviz = "^0.0.2"
matplotlib = "^3.8.0"
tqdm = "^4.66.1"
hydra-core = "^1.3.2"
ray = {extras = ["data", "rllib", "serve", "train", "tune"], version = "^2.8.0"}
omegaconf = "^2.3.0"
hydra-colorlog = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
force-cuda11 = "python -m pip install torch==1.12.0+cu113 torchvision==0.13.0+cu113 -f https://download.pytorch.org/whl/torch_stable.html"
